% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fy2}
\alias{fy2}
\title{Some fake Food & You 2 data.}
\format{
A data frame with 100 rows and 6 variables:
\describe{
\item{id}{ID}
\item{age}{Age, in years}
\item{sex}{Sex}
\item{marstat}{Marital status: Married/civil partnership; Single: living as a couple; Single: not living as a couple; Divorced/Widowed/Separated}
\item{urban}{Urban or rual dweller}
\item{food_sec}{Food security: High; Marginal; Low; Very low}
}
}
\usage{
fy2
}
\description{
A fake dataset containing some simulated demographic data for FY2.
}
\keyword{datasets}
